// src/modules/contact/contact.interface.ts
export interface IContact {
  id: string
  name: string
  cpf: string
  phone: string
  zipCode: string
  street: string
  number: string
  complement: string
  neighborhood: string
  city: string
  state: string
  region: string
  userId: string
  lat: number
  lng: number
  createdAt: string
  updatedAt: string
}